refactor(creator): move Undo/Redo into the canvas toolbar, add icons
Follow-up to the undo/redo stack: Undo/Redo now live in the canvas
toolbar (Panel, top-center) next to Auto-layout/Group Selection --
editing-history actions belong with the other canvas-editing controls,
not the ScenarioBar's file-lifecycle actions (New/Save/Export). Also
swap the plain unicode arrows for stroke-icon SVGs matching the app's
existing icon convention (the focus-mode button in this same toolbar):
Lucide's rotate-ccw/rotate-cw, a circular arrow with a hooked
arrowhead.

Verified with a chrome-devtools text snapshot: Undo/Redo render in the
canvas toolbar, are gone from the Scenario section, and the console is
clean (only the pre-existing favicon 404).

apps/creator is private; no changeset needed.
